requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3

asttokens==2.0.8
astunparse==1.6.3

av==9.2.0

black==22.3.0

brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7

charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0

cycler==0.11.0
Cython==0.29.32

executing==0.10.0
fairscale==0.4.8

flatbuffers==1.12
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
imageio==2.21.1
imageio-ffmpeg==0.4.7
imgaug==0.2.6
importlib-metadata==4.12.0
iopath==0.1.9

joblib==1.1.0

keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lucid==0.2.3

Markdown==3.4.1

matplotlib==3.5.3

mediapipe==0.8.10.1

moviepy==1.0.3
mypy-extensions==0.4.3

networkx==2.8.6
ninja==1.10.2.3

numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0

pandas==1.5.2

pathspec==0.9.0

Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
portalocker==2.5.1
proglog==0.1.10

prompt-toolkit==3.0.30
protobuf==3.19.4

pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4

pydot==1.4.2
Pygments==2.13.0
PyOpenGL==3.1.6

PyQt5-sip==12.11.0
pyrealsense2==2.53.1.4623

PyWavelets==1.3.0
PyYAML==6.0

requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0

stack-data==0.4.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0

threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7

torch==1.12.1+cu117
torchaudio==0.12.1+cu117
torchvision==0.13.1+cu117

tqdm==4.64.0
traitlets==5.3.0

webencodings==0.5.1

Werkzeug==2.2.2

wrapt==1.14.1
yacs==0.1.8
zipp==3.8.1
